+++ gui_widgets/sbg_huds.gd
@@ -33,7 +33,9 @@ extends GridContainer
 const NULL_COLOR := Color.BLACK
 
 
-var enable_wiki: bool = IVCoreSettings.enable_wiki
+## Allows row lables to be wiki links, but only if [IVWikiManager] exists and
+## [method IVWikiManager.has_page] returns true.
+@export var allow_wiki_links := false
 
 var has_headers := true
 var column_master: Control # if set, column widths follow master children
@@ -50,15 +52,16 @@ var headers: Array[StringName] = [&"LABEL_POINTS", &"LABEL_ORBITS"]
 var header_hints: Array[StringName] = [&"HINT_POINTS_CKBX_COLOR", &"HINT_ORBITS_CKBX_COLOR"]
 
 
-var _wiki_titles: Dictionary[StringName, String] = IVTableData.wiki_lookup
 var _points_ckbxs: Array[CheckBox] = []
 var _orbits_ckbxs: Array[CheckBox] = []
 var _points_color_pkrs: Array[ColorPickerButton] = []
 var _orbits_color_pkrs: Array[ColorPickerButton] = []
 var _suppress_update := false
 
-@onready var _sbg_huds_state: IVSBGHUDsState = IVGlobal.program[&"SBGHUDsState"]
+
 @onready var _n_rows := rows.size()
+@onready var _sbg_huds_state: IVSBGHUDsState = IVGlobal.program[&"SBGHUDsState"]
+@onready var _wiki_manager: IVWikiManager = IVGlobal.program.get(&"WikiManager")
 
 
 
@@ -87,20 +90,19 @@ func _ready() -> void:
 		var row_name: StringName = rows[i][0]
 		var groups: Array[StringName] = rows[i][1]
 		var is_indent: bool = rows[i][2]
+		var row_indent := indent if is_indent else ""
 		# row label
-		if enable_wiki and _wiki_titles.has(row_name):
+		if allow_wiki_links and _wiki_manager and _wiki_manager.has_page(row_name):
 			var rtlabel := RichTextLabel.new()
-			rtlabel.meta_clicked.connect(_on_meta_clicked.bind(row_name))
+			rtlabel.meta_clicked.connect(_on_meta_clicked)
 			rtlabel.bbcode_enabled = true
 			rtlabel.fit_content = true
 			rtlabel.scroll_active = false
-			if is_indent:
-				rtlabel.text = indent
-			rtlabel.text += "[url]" + tr(row_name) + "[/url]"
+			rtlabel.parse_bbcode('%s[url="%s"]%s[/url]' % [row_indent, row_name, tr(row_name)])
 			add_child(rtlabel)
 		else:
 			var label := Label.new()
-			label.text = indent + tr(row_name) if is_indent else tr(row_name)
+			label.text = row_indent + tr(row_name)
 			add_child(label)
 		# points
 		var hbox := HBoxContainer.new()
@@ -295,9 +297,8 @@ func _resize_columns_to_en_width(delay_frames := 0) -> void:
 		child.size.x = min_width
 
 
-func _on_meta_clicked(_meta: String, row_name: StringName) -> void:
-	var wiki_title: String = _wiki_titles[row_name]
-	IVGlobal.open_wiki_requested.emit(wiki_title)
+func _on_meta_clicked(meta: String) -> void:
+	_wiki_manager.open_page(meta)
 
 
 func _settings_listener(setting: StringName, _value: Variant) -> void: